requirements.txt
Flask==1.1.2
Werkzeug==0.16.1
requests==2.22.0
mysql_connector_repackaged==0.3.1
PyJWT==2.0.1
mysql-connector-python==8.0.21
